Reverted the enum underlying types added by the clang-tidy work
The underlying type of an enum is a design decision, so it is no longer added automatically.

- Reverted 0ebe9490 (performance-enum-size outside FslBase) and the enum part of d4161f00 (FslBase),
  including the <cstdint> includes that were only added for them.
- The new FramePacing enums (FramePacingMarkerKind, FramePacingMarkerSlot and FramePacingRunState)
  no longer specify an underlying type either.
- performance-enum-size is disabled in .clang-tidy and its EnumIgnoreList option was removed.

All enum declarations now match 60952802 (the state before this work) apart from the enums that are
new since then, which use the default underlying type.

+++ DemoApps/Shared/AntiAliasing/include/Shared/AntiAliasing/SupportedSampleCountFlags.hpp
@@ -33,11 +33,9 @@
 
 #include <FslBase/BasicTypes.hpp>
 
-#include <cstdint>
-
 namespace Fsl
 {
-  enum class SupportedSampleCountFlags : uint8_t
+  enum class SupportedSampleCountFlags : uint32_t
   {
     NoFlags = 0,
     Samples1 = 1,
