+++ DemoFramework/FslSimpleUI/App/include/FslSimpleUI/App/UIAppResourceFlag.hpp
@@ -33,12 +33,10 @@
 
 #include <FslBase/BasicTypes.hpp>
 
-#include <cstdint>
-
 namespace Fsl
 {
   //! Represents the operations that was performed
-  enum class UIAppResourceFlag : uint16_t
+  enum class UIAppResourceFlag : uint32_t
   {
     //! A empty flag
     Undefined = 0,
