+++ DemoFramework/FslSimpleUI/Base/include/FslSimpleUI/Base/Gesture/GestureAxis.hpp
@@ -33,11 +33,9 @@
 
 #include <FslBase/BasicTypes.hpp>
 
-#include <cstdint>
-
 namespace Fsl::UI
 {
-  enum class GestureAxis : uint8_t
+  enum class GestureAxis : uint32_t
   {
     // A empty flag
     NotDefined = 0x00,
