+++ DemoFramework/FslSimpleUI/Render/Base/include/FslSimpleUI/Render/Base/Command/EncodedCommandStateFlags.hpp
@@ -32,11 +32,10 @@
  ****************************************************************************************************************************************************/
 
 #include <FslBase/BasicTypes.hpp>
-#include <cstdint>
 
 namespace Fsl::UI
 {
-  enum class EncodedCommandStateFlags : uint16_t
+  enum class EncodedCommandStateFlags : uint32_t
   {
     // Can't be called none due to collision with X11 files
     Nothing = 0,
